refactor: Context의 책임을 분리하여 개선
